Added "might_have_changes" optimization to importer
